% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_aux.R
\name{block.Sig}
\alias{block.Sig}
\title{Creates block sigma}
\usage{
block.Sig(pars, coords, reg.ind, cov.model)
}
\arguments{
\item{pars}{Log covariance parameters for the regionalized kriging model.
See \code{\link{rlikfit}}.}

\item{coords}{Coordinates of the locations}

\item{reg.ind}{Vector of region indicators}

\item{cov.model}{Covariance models for each region.}
}
\description{
To do.
}
\details{
Details to be added.
}
\author{
Paul Sampson, Joshua Keller
}

