% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kriging.R
\name{predict.rlikfit}
\alias{predict.rlikfit}
\title{Predict from a regionalized kriging model}
\usage{
\method{predict}{rlikfit}(object, X.mon, coords.mon, reg.mon, y, X.pred,
  coords.pred, reg.pred, ...)
}
\arguments{
\item{object}{An object of class \code{\link{rlikfit}} from which to compute predictions.}

\item{X.mon}{Kriging covariates for monitors}

\item{coords.mon}{Monitor coordinates}

\item{reg.mon}{Vector of regions for monitors}

\item{y}{Monitor pollutant concentrations}

\item{X.pred}{Covariates (i.e., PLS scores) for prediction locations}

\item{coords.pred}{Prediction coordinates}

\item{reg.pred}{Vector of regions for prediction sites}

\item{...}{ignored additional arguments}
}
\description{
Makes conditional expectations (kriging predictions) from
	a regionalized kriging model and computes prediction variances.
}
\author{
Joshua Keller, Paul Sampson
}
\seealso{
rlikfit
}

