% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kriging.R
\name{print.rlikfit}
\alias{print.rlikfit}
\title{Print details for class \code{rlikfit}}
\usage{
\method{print}{rlikfit}(x, ...)
}
\arguments{
\item{x}{object of class \code{rlikfit}}

\item{...}{Ignored additional arguments.}
}
\description{
\code{\link[base:print]{print}} method for class \code{rlikfit}.
}

