% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kriging.R
\name{prof.rlik}
\alias{prof.rlik}
\title{Evaluates profile likelihood function for (regionalized) universal kriging.}
\usage{
prof.rlik(x, R, coords, reg.ind, y, useSTpackage = TRUE)
}
\arguments{
\item{x}{Log covariance parameters}

\item{R}{Kriging covariate model matrix}

\item{coords}{Coordinates}

\item{reg.ind}{Region indicator}

\item{y}{Observed outcome values.}

\item{useSTpackage}{See details.}
}
\description{
To do.
}
\details{
If the SpatioTemporal package is available, uses block cholesky
		functions within that package.
}
\author{
Paul Sampson, Josh Keller
}
\seealso{
\link{rlikfit}
}

