% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kriging.R
\name{summary.rlikfit}
\alias{summary.rlikfit}
\title{Compute summary details for class \code{rlikfit}}
\usage{
\method{summary}{rlikfit}(object, ...)
}
\arguments{
\item{object}{object of class \code{rlikfit}}

\item{...}{Ignored additional arguments.}
}
\description{
\code{\link[base:print]{summary}} method for class \code{rlikfit}.
}

