% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_aux.R
\name{varcov.eff}
\alias{varcov.eff}
\title{Creates exponential covariance matrix}
\usage{
varcov.eff(coords, pars)
}
\arguments{
\item{coords}{Coordiantes}

\item{pars}{A vector of three log covariance parameters: nugget, sill, and range.}
}
\value{
Returns an \eqn{n x n} matrix.
}
\description{
Creates a covariance matrix according to an expontial model.
}
\author{
Paul Sampson
}

