% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_aux.R
\name{block.Sig}
\alias{block.Sig}
\title{Creates block sigma}
\usage{
block.Sig(l.pars, coords, reg.ind)
}
\arguments{
\item{l.pars}{Log covariance parameters}

\item{coords}{Coordiantes}

\item{reg.ind}{Vector of region indicators}
}
\description{
To do.
}
\details{
Details to be added.
}
\author{
Paul Sampson, Josh Keller
}

