% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_aux.R
\name{inf.beta}
\alias{inf.beta}
\title{Infers Beta from maximized covariance pars.}
\usage{
inf.beta(l.pars, X, coords, reg.ind, y, Sig = NULL, useSTpackage = TRUE)
}
\arguments{
\item{l.pars}{Log covariance parameters}

\item{X}{kriging covariates}

\item{coords}{Coordiantes}

\item{reg.ind}{Vector of region indicators}

\item{y}{outcome variable}

\item{Sig}{block-diagonal covariance matrix. Default value is \code{NULL}, and the
matrix is computed by a call to \code{\link{block.Sig}}. 
If it is already computed, providing it can speed up code.}

\item{useSTpackage}{See Details.}
}
\description{
Computes estimates of regression coefficients from
	covariance estimates.
}
\details{
This computes (X'SinvX)^(-1)(X'Sinv*y). Takes advantage of 
	functions in SpatioTemporal package if available.
}
\author{
Paul Sampson, Casey Olives, Joshua Keller
}

