% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kriging.R
\name{prof.rlik}
\alias{prof.rlik}
\title{Profile Likelihood for Kriging}
\usage{
prof.rlik(pars, X, coords, reg.ind, y, cov.model = "exp",
  useSTpackage = TRUE)
}
\arguments{
\item{pars}{Log covariance parameters}

\item{X}{Kriging covariate model matrix}

\item{coords}{Coordinates}

\item{reg.ind}{Region indicator}

\item{y}{Observed outcome values.}

\item{cov.model}{Covariance model for each region.}

\item{useSTpackage}{Logical. If the SpatioTemporal package is available, 
uses block cholesky functions within that package for
improved performance.}
}
\value{
Scalar value giving the *negative* log profile likelihood
	for the kriging model.
}
\description{
Evaluates profile likelihood function for (regionalized) universal kriging model.
}
\details{
This function is primarily intended to be called
	via \code{\link{rlikfit}}, and thus has limited 
	checking of arguments.
}
\author{
Paul Sampson, Josh Keller
}
\seealso{
\link{rlikfit}
}

