% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kriging.R
\name{rlikfit}
\alias{rlikfit}
\title{Likelihood Fit for Regoinalized Universal Kriging Model}
\usage{
rlikfit(pars, y, X, coords, reg.ind, hess = TRUE, optim.args = NULL)
}
\arguments{
\item{pars}{Vector of initial starting values for log-covariance parameters. 
Should have length \eqn{3r}, where    
where \eqn{r} is the number of 
distinct regions. Values should follow the order of nugget (\eqn{\tau}),
sill (\eqn{\sigma}), and range (\eqn{\phi}). 
e.g. c(tau1,sigma1,rho1,tau2,sigma2,rho2,tau3,sigma3,rho3)}

\item{y}{Dependent variable to be modeled}

\item{X}{\eqn{n x d} matrix of kriging covariates for the mean component}

\item{coords}{\eqn{N x 2} matrix of coordinates}

\item{reg.ind}{\eqn{N}-vector of distinct kriging regions.  Defaults to  \code{reg.ind = rep(1,N)}, 
which yields traditional universal kriging.}

\item{hess}{Logical; whether or not to compute hessian and test whether it is positive-definite.}

\item{optim.args}{List of arguments passed to \code{\link{optim}}. See details.}
}
\value{
An object of class 'rlikfit' containing:
\item{log.cov.pars}{maximized log-covariance parameters}
 \item{beta}{kriging regression covariate estimates}
 \item{max.log.lik}{maximized log-likelihood}
 \item{hess.pd}{ indicator of positive definite hessian, if hess=TRUE}
 \item{r}{ Number of regions}
}
\description{
Maximizes a kriging likelihood that can include a regionalized variogram and regionalized coefficients.
}
\details{
The maximization is done via profile likelihood, using the function \code{\link{prof.rlik}}.

 The default upper bound for the nugget and sill is log(10), and the default lower bound is log(0.0001). Default bounds for the range are log(0.001) and log(5000). FINISH
}
\author{
Joshua Keller, Paul Sampson
}

